+++ pkg/apis/softwarecomposition/networkpolicy/v1/networkpolicy_test.go
