+++ pkg/apis/softwarecomposition/v1beta1/networkpolicy/networkpolicy.go
@@ -2,7 +2,7 @@ package networkpolicy
 
 import (
 	sc "github.com/kubescape/storage/pkg/apis/softwarecomposition"
-	np "github.com/kubescape/storage/pkg/apis/softwarecomposition/networkpolicy"
+	np "github.com/kubescape/storage/pkg/apis/softwarecomposition/networkpolicy/v1"
 	"github.com/kubescape/storage/pkg/apis/softwarecomposition/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
